+++ R/fitLogistic.R
@@ -84,11 +84,11 @@ fitLogistic <- function(endpoint, placebo, data, alternative, ...) {
   ret <- NULL
 
   for(trt_arm in treatment_arms){
-    sub_data <- filtered_data %>% dplyr::filter(arm %in% c(placebo, trt_arm))
+    sub_data <- filtered_data %>% dplyr::filter(.data$arm %in% c(placebo, trt_arm))
 
     # Fit the logistic regression model
     fit <- summary(glm(as.formula(formula_str),
-                       family = binomial,
+                       family = 'binomial',
                        data = sub_data)
                    )$coef