+++ R/fitLogrank.R
@@ -70,7 +70,7 @@ fitLogrank <- function(endpoint, placebo, data, ...) {
   ret <- NULL
 
   for(trt_arm in treatment_arms){
-    sub_data <- filtered_data %>% dplyr::filter(arm %in% c(placebo, trt_arm))
+    sub_data <- filtered_data %>% dplyr::filter(.data$arm %in% c(placebo, trt_arm))
 
     # Create the formula
     formula_str <- paste0("Surv(time = ", endpoint,